backend/requirements.txt
# backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-telegram-bot==21.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
redis==5.0.1
PyMuPDF==1.23.8
python-docx==1.1.0
httpx
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
Pillow==10.2.0